requirements.txt
comet_ml==3.32.8
jpeg4py==0.1.4
loguru
matplotlib
numpy
opencv_python
chardet
Pillow
pyrender
pytorch_lightning==2.0.0
scipy
smplx==0.1.28
tqdm
trimesh==3.9.21
scikit-image
imgui
aitviewer==1.8.1
chumpy
black
autopep8
flake8
pylint
isort
easydict
pygit2
ipdb
opencv-python-headless
